Convert type from double to float

+++ optimizations/JPEG/Images/Pixel.cs
@@ -7,14 +7,14 @@ public readonly struct Pixel
 {
     // Переделать в byte
 
-    public double R { get; }
-    public double G { get; }
-    public double B { get; }
-    public double Y { get; }
-    public double Cb { get; }
-    public double Cr { get; }
+    public float R { get; }
+    public float G { get; }
+    public float B { get; }
+    public float Y { get; }
+    public float Cb { get; }
+    public float Cr { get; }
     
-    public Pixel(double firstComponent, double secondComponent, double thirdComponent, PixelFormat pixelFormat)
+    public Pixel(float firstComponent, float secondComponent, float thirdComponent, PixelFormat pixelFormat)
     {
         if (!new[] { PixelFormat.RGB, PixelFormat.YCbCr }.Contains(pixelFormat))
             throw new FormatException("Unknown pixel format: " + pixelFormat);
@@ -25,9 +25,9 @@ public readonly struct Pixel
             G = secondComponent;
             B = thirdComponent;
 
-            Y = 16.0 + (65.738 * R + 129.057 * G + 24.064 * B) / 256.0;
-            Cb = 128.0 + (-37.945 * R - 74.494 * G + 112.439 * B) / 256.0;
-            Cr = 128.0 + (112.439 * R - 94.154 * G - 18.285 * B) / 256.0;
+            Y = 16.0f + (65.738f * R + 129.057f * G + 24.064f * B) / 256.0f;
+            Cb = 128.0f + (-37.945f * R - 74.494f * G + 112.439f * B) / 256.0f;
+            Cr = 128.0f + (112.439f * R - 94.154f * G - 18.285f * B) / 256.0f;
         }
 
         if (pixelFormat == PixelFormat.YCbCr)
@@ -36,9 +36,9 @@ public readonly struct Pixel
             Cb = secondComponent;
             Cr = thirdComponent;
 
-            R = (298.082 * Y + 408.583 * Cr) / 256.0 - 222.921;
-            G = (298.082 * Y - 100.291 * Cb - 208.120 * Cr) / 256.0 + 135.576;
-            B = (298.082 * Y + 516.412 * Cb) / 256.0 - 276.836;
+            R = (298.082f * Y + 408.583f * Cr) / 256.0f - 222.921f;
+            G = (298.082f * Y - 100.291f * Cb - 208.120f * Cr) / 256.0f + 135.576f;
+            B = (298.082f * Y + 516.412f * Cb) / 256.0f - 276.836f;
         }
     }
 }
